Smart default priors for all terms is now a thing
It seems to work great. See the new examples/default_priors.ipynb notebook
for a little demonstration. Note that there curently seems to be an issue
with adding random slopes for categorical predictors with >2 levels (as
documented in the notebook), and once this is resolved, the corresponding
code that handles default priors for that case will likely need to be revised.