Slope-only and cell-means models now pass tests
The tests now take a lot longer (about 9 mins on my laptop) because they
now fit the models (taking 1 sample) in addition to building them. I think
this is pretty much unavoidable because a lot of the errors don't crop up
until you try to build the model.

The cell-means-with-covariates model currently fails because the default
prior on the categorical predictor has infinite variance. This happens because
it tries to treat that case like a normal case, when in reality it's arguably
an edge case. The slope_constant term that goes into the default priors
contains a 1 - r2x term in the denominator, where r2x is the R^2 from
regressing that predictor on all the other predictors. In this model, r2x
is 1, because this auxiliary regression includes an intercept even though
the main regression model has no intercept. So this makes the denominator 0,
and the slope_constant term becomes infinity. Modifying that auxiliary
regression is one possible way forward (maybe), but really what we want is to
treat this case pretty much like the cell-means case (w/ no covariate).