Tweak SD(Y) in binomial models for more sensible default priors
This commit introduces an interim solution where SD(Y) is changed for
binomial/logit and binomial/probit models (for when the latter are
implemented), but there is a more general problem about how to define
default priors in a sensible way for various non-normal models. I'm about
to write an issue with some thoughts.