Slight tweak to GLMM example in README
I think this example makes more sense with random intercepts for schools
instead of random intercepts for students. This way, each row is a student.
The problem with the earlier setup (where each row must be something like a
class that student took) is that each student presumably has a single value
for 'graduate', so all of the student effects would be either Inf (if the
student graduated) or -Inf (if the student didnt' graduate). This reflects a
more general issue with fitting GLMMs when the outcome varies at "level 2."