Changed default SD priors from HalfCauchy to HalfNormal
The defaults are pretty wide either way, but HalfNormal is a little less
forgiving of crazy extreme draws. Based on some small-scale tests on the
crossed_random dataset (which has some issues in getting good estimates
of the random effect variances), this change does seem to have a noticeably
beneficial effect in terms of the chains being a little better behaved.