ModelResults.summary output is mo concise, mo infomative, mo betta
It is now based on pm.df_summary() rather than pm.summary(). Terms with
multiple levels have the actual level names printed in the output.
Internally transformed variables are suppressed by default. Random
effects are not summarized by default (but their SDs are). Between this
and e02a276, this effectively closes #36 and definitely closes #41.

Also adds test both for this and for the traceplots and priors plots.
The tests are in test_model basically because we have to fit a model
before we can test the fitted model summary functions.