ModelResults.get_trace() returns the samples in nice, labeled DF
The main advantage of this over working directly with the MultiTrace
is that this nicely labels all levels of all terms. Arguably there is
also an advantage to having the samples in a DataFrame object that
will be familiar to users. Also this is an abstract method that can in
principle be used with other backends.