Modified test to draw enough samples to compute `mc_error`
The model in which we test the summary function was based on just 10
samples, but this was not enough to compute `mc_error`, which was
always a column of NaN values. Now we draw 100 samples, which seems
to be about the smallest number that is large enough to fill in the
`mc_error` column.