Line comment tweaks + updated example notebooks
I implemented all of the changes suggested in the line comments, except
for moving the tests of the summary/plotting functions to a separate
testing module (which I agree we should probably do). I also updated the
ESCS_multiple_regression notebook to take advantage of the new featuers,
and added the ANES_logistic_regression notebook.