Fixed the test of _filter_names
It looks like some versions of pymc3 add a trailing '_' to transformed
vars and some don't. Mine doesn't but it seems that the travis version
does. So for consistency across pymc3 versions, the test strips off
any trailing '_' found in the var names. Note that this is purely a
test issue and actual user functionality should be fine either way.