Fix missing 'family' argument to sm.GLM()
One of the calls to sm.GLM() in the default prior code was missing the
family specification, so the results were not right for non-Normal models.
I really do not know how the tests were passing with this argument missing,
since the logistic regression test was being fit to float values. And
apparently it was by pure chance that the results for the ANES logistic
example were still sensible despite this problem. Anyway, fixed now.