Adjust points at which loglike function is evaluated
This results in a slightly more accurate approximation to the log likelihood.
Furthermore, now approximation errors tend to be in the direction of making
the default priors a bit too wide rather than a bit too narrow, which is
probably preferable. Accordingly, the default prior width for non-Normal
models has been changed back from 'superwide' to 'wide.'