Convergence diagnostics added when njobs > 1; README updated
When a model is run with njobs > 1, the DataFrame returned by
ModelResults.summary() now includes columns giving the effective sample
size and Gelman-Rubin statistics for each parameter. If the model is run
with njobs = 1, a warning is displayed saying that multiple chains are
needed in order to compute convergence diagnostics.

Also updates the README section discussing the default priors, and includes
a link to the new technical paper on arxiv.