Fix corner case in prior assignment
There was a problem in the admittedly unusual case where the fixed effects
had cell-means coding, but the random effects for that same factor had
k-1 dummy coding (e.g., 0+threecats in the fixed effects, 1+threecats in the
random effects). This was actually a problem before too, but the prior was
just being silently handled in a not-quite-appropriate way. Now it is
handled appropriately.

The remaining failing tests should be mostly related to consistency between
pymc3 and stan.