Change 'Binomial' to 'Bernoulli' (will add binomial soon)
We currently handle logistic regression, and although we call the
distribution/family binomial, we actually only correctly handle the case
where n_trials = 1, i.e., the bernoulli case. Down the road it would be
best to separately handle the bernoulli and the general binomial cases.
So for now, rename all binomial references to bernoulli, and when we
go in and add a bunch of other distributions, we'll make sure to add
binomial along with those.