Ordering of variables in plot()
Following suggestion #114 Added a simple re-ordering code to produce an easier to read plot output. Order is defined as follow : Intercept, main effects, interactions, random effects and sd. Within a subcategory all names are ordered with length of the name.
Would be nice to apply to the summary output but I did not find a quick way to do it.